Improved Number Handwriting Model
Classic NMIST number HW model--- Improved HW models made several different ways for learning high-accuracy model training. I also include a really simple and effective frontend that is universal and can be made into n app for any platform very easily. We also improved DRAMATICALLY upon the teacher-led models, and we show how to get the frontend to create a visual and essentially let you "see what is sees" so to speak. we have a unique build here for resizing and centering the image because we were receiving distortions and the model shrunk the image down into a blobby mess(example output of a perfectly drawn 5 turned to mush included from original frontend design).

So now that the model is super accurate, fast to train and the interface ceters and appropriately resizes the digit with a thinner stroke, we have an essentially flawless machine than not only can predict accurately a drawn number but can usually predict a partially drawn digit with impressive accuracy